2nd Attempt at first commit
UK shape files were too large to be uploaded, and even using git's LFS system and `git push -u origin main --force` didn't work as the shape files were over 100mb. Extracted the Scotland data from it and converted it to a geojson, and moved the uk spatial data to the `raw_data` folder, which is in the .gitignore.

Scotland geojson is now <100mb.